@@ -1,13 +1,13 @@
 class CupertinoOptions {
-  final String backgroundColor;
-  final String doneButtonTitle;
-  final String selectionShadowColor;
-  final String selectionStrokeColor;
-  final String selectionFillColor;
-  final String selectionTextColor;
-  final String selectionCharacter;
-  final String takePhotoIcon;
-  final bool autoCloseOnSelectionLimit;
+  final String? backgroundColor;
+  final String? doneButtonTitle;
+  final String? selectionShadowColor;
+  final String? selectionStrokeColor;
+  final String? selectionFillColor;
+  final String? selectionTextColor;
+  final String? selectionCharacter;
+  final String? takePhotoIcon;
+  final bool? autoCloseOnSelectionLimit;
 
   const CupertinoOptions({
     this.backgroundColor,
