@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:io' show Platform;
 
 import 'package:flutter/services.dart';
-import 'package:meta/meta.dart';
 import 'package:multi_image_picker2/multi_image_picker2.dart';
 import 'package:multi_image_picker2/src/exceptions.dart';
 
@@ -34,15 +33,13 @@ class MultiImagePicker {
   /// penalty. How to request the original image or a thumb
   /// you can refer to the docs for the Asset class.
   static Future<List<Asset>> pickImages({
-    @required int maxImages,
+    required int maxImages,
     bool enableCamera = false,
     List<Asset> selectedAssets = const [],
     CupertinoOptions cupertinoOptions = const CupertinoOptions(),
     MaterialOptions materialOptions = const MaterialOptions(),
   }) async {
-    assert(maxImages != null);
-
-    if (maxImages != null && maxImages < 0) {
+    if (maxImages < 0) {
       throw new ArgumentError.value(maxImages, 'maxImages cannot be negative');
     }
 
@@ -61,7 +58,8 @@ class MultiImagePicker {
               .toList(),
         },
       );
-      var assets = List<Asset>();
+
+      var assets = <Asset>[];
       for (var item in images) {
         var asset = Asset(
           item['identifier'],
@@ -75,7 +73,11 @@ class MultiImagePicker {
     } on PlatformException catch (e) {
       switch (e.code) {
         case "CANCELLED":
-          throw NoImagesSelectedException(e.message);
+          throw NoImagesSelectedException(e.message!);
+        case "PERMISSION_DENIED":
+          throw PermissionDeniedException(e.message!);
+        case "PERMISSION_PERMANENTLY_DENIED":
+          throw PermissionPermanentlyDeniedExeption(e.message!);
         default:
           throw e;
       }
@@ -90,17 +92,13 @@ class MultiImagePicker {
   /// refer to [Asset] class docs.
   ///
   /// The actual image data is sent via BinaryChannel.
-  static Future<bool> requestThumbnail(
-      String identifier, int width, int height, int quality) async {
-    assert(identifier != null);
-    assert(width != null);
-    assert(height != null);
-
-    if (width != null && width < 0) {
+  static Future<bool?> requestThumbnail(
+      String? identifier, int width, int height, int quality) async {
+    if (width < 0) {
       throw new ArgumentError.value(width, 'width cannot be negative');
     }
 
-    if (height != null && height < 0) {
+    if (height < 0) {
       throw new ArgumentError.value(height, 'height cannot be negative');
     }
 
@@ -110,7 +108,7 @@ class MultiImagePicker {
     }
 
     try {
-      bool ret = await _channel.invokeMethod(
+      bool? ret = await _channel.invokeMethod(
           "requestThumbnail", <String, dynamic>{
         "identifier": identifier,
         "width": width,
@@ -121,11 +119,11 @@ class MultiImagePicker {
     } on PlatformException catch (e) {
       switch (e.code) {
         case "ASSET_DOES_NOT_EXIST":
-          throw AssetNotFoundException(e.message);
+          throw AssetNotFoundException(e.message!);
         case "PERMISSION_DENIED":
-          throw PermissionDeniedException(e.message);
+          throw PermissionDeniedException(e.message!);
         case "PERMISSION_PERMANENTLY_DENIED":
-          throw PermissionPermanentlyDeniedExeption(e.message);
+          throw PermissionPermanentlyDeniedExeption(e.message!);
         default:
           throw e;
       }
@@ -140,9 +138,9 @@ class MultiImagePicker {
   /// refer to [Asset] class docs.
   ///
   /// The actual image data is sent via BinaryChannel.
-  static Future<bool> requestOriginal(String identifier, quality) async {
+  static Future<bool?> requestOriginal(String? identifier, quality) async {
     try {
-      bool ret =
+      bool? ret =
           await _channel.invokeMethod("requestOriginal", <String, dynamic>{
         "identifier": identifier,
         "quality": quality,
@@ -151,7 +149,7 @@ class MultiImagePicker {
     } on PlatformException catch (e) {
       switch (e.code) {
         case "ASSET_DOES_NOT_EXIST":
-          throw AssetNotFoundException(e.message);
+          throw AssetNotFoundException(e.message!);
         default:
           throw e;
       }
@@ -159,13 +157,13 @@ class MultiImagePicker {
   }
 
   // Requests image metadata for a given [identifier]
-  static Future<Metadata> requestMetadata(String identifier) async {
-    Map<dynamic, dynamic> map = await _channel.invokeMethod(
+  static Future<Metadata> requestMetadata(String? identifier) async {
+    Map<dynamic, dynamic> map = await (_channel.invokeMethod(
       "requestMetadata",
       <String, dynamic>{
         "identifier": identifier,
       },
-    );
+    ) as FutureOr<Map<dynamic, dynamic>>);
 
     Map<String, dynamic> metadata = Map<String, dynamic>.from(map);
     if (Platform.isIOS) {
@@ -198,6 +196,6 @@ class MultiImagePicker {
       }
     });
 
-    return map;
+    return map as Map<String, dynamic>;
   }
 }
