@@ -244,9 +244,9 @@ public class MultiImagePickerPlugin implements
 
         @Override
         protected void onPostExecute(ByteBuffer buffer) {
-            super.onPostExecute(buffer);
-            this.messenger.send("multi_image_picker/image/" + this.identifier + ".original", buffer);
+            super.onPostExecute(buffer);     
             if(buffer != null) {
+                this.messenger.send("multi_image_picker/image/" + this.identifier + ".original", buffer);
                 buffer.clear();
             }
         }
@@ -619,7 +619,9 @@ public class MultiImagePickerPlugin implements
             List<HashMap<String, Object>> result = new ArrayList<>(photos.size());
             for (Uri uri : photos) {
                 if(uri == null){
-                    continue;
+                    clearMethodCallAndResult();
+                    return false;
+                  //  continue;
                 }
                 HashMap<String, Object> map = new HashMap<>();
                 map.put("identifier", uri.toString());
